Move the software client from the CLI to the service
